EMPIREDB-103 remaining changes
git-svn-id: https://svn.apache.org/repos/asf/incubator/empire-db/trunk@1180733 13f79535-47bb-0310-9956-ffa450edef68

+++ empire-db-codegen/src/main/java/org/apache/empire/db/codegen/CodeGenParser.java
@@ -432,7 +432,7 @@ public class CodeGenParser {
 		case Types.DOUBLE:
 		case Types.FLOAT:
 		case Types.REAL:
-			empireType = DataType.DOUBLE;
+			empireType = DataType.FLOAT;
 			break;
 		case Types.DECIMAL:
 		case Types.NUMERIC:
